+++ lib/json_api_ruby.rb
@@ -3,6 +3,7 @@ end
 
 require_relative 'json_api_ruby/error_resource'
 require_relative 'json_api_ruby/serializer'
+require_relative 'json_api_ruby/association'
 require_relative 'json_api_ruby/resource'
 require_relative 'json_api_ruby/exceptions'
 require_relative 'json_api_ruby/configuration'
