+++ lib/json_api_ruby/includes.rb
@@ -3,7 +3,7 @@ module JsonApi
   class Includes
     attr_reader :name
     attr_writer :includes
-    attr_accessor :next
+    attr_writer :next
 
     # Returns:
     #   #<Includes>
@@ -20,17 +20,27 @@ module JsonApi
       zipped = first_array.zip(*other_arrays)
       first_array = zipped.shift
       first_include.includes = first_array.uniq.compact
+
       zipped.inject(first_include) do |base_include, object_name|
         new_include = self.new
         new_include.includes = object_name.uniq.compact
         base_include.next = new_include
         new_include
       end
+
       first_include
+    end
+
+    def has_name?(name)
+      includes.include?(name)
     end
 
     def includes
       @includes || []
+    end
+
+    def next
+      @next || self.class.new
     end
   end
 end
